scheduled-injest/requirements.txt
requests
python-dotenv
schedule
caldav
psycopg2
playwright
asyncio
psutil
GPUtil
fastembed
numpy